nilm-chat-backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
pydantic>=2.3.0
python-dotenv>=1.0.0
asyncpg>=0.28.0  # For PostgreSQL
pandas>=2.1.0
python-multipart>=0.0.6
httpx>=0.24.1

# Huggingface and Torch requirements
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99